Add support for compiling libcrypto with gcc -std=gnu99
Add support for compiling with non-C11 gcc compiler.
In order to do so, configure with:
    $ cmake -DFORCE_C_STD_FLAGS=-std=gnu99 .
    $ make crypto

This forces compilation with gnu99 extensions for the C code.
Note that this support does not include compilation of C++11
code, hence tests can't be compiled this way.
Thus, in order to compile libcrypto only, do 'make crypto'
instead of 'make all'.

Changes are limited to:
- cmake support for FORCE_C_STD_FLAGS
- definition of alignof/alignas with GCC extensions when activated
  instead of stdalign.h (C11) inclusion
- C11 constructs (unnamed union/struct) are handle though gnu99
  without any change

In order to test in this mode, use a C/C++11 compiler and run tests
with:
    $ cmake -DFORCE_C_STD_FLAGS=-std=gnu99 .
    $ make -j $(nproc) all
    $ make run_tests
    ...
    All tests passed!
    0/0/2575/2575/2575
    PASS
    [100%] Built target run_tests

Tested with:
- build crypto: Ubuntu 12.04 (gcc 4.6.2) & CentOS 6 (gcc 4.4.7):
    $ cmake -DOPENSSL_NO_ASM=1 -DFORCE_C_STD_FLAGS=-std=gnu99 .
    $ make crypto
    [100%] Built target crypto
- build all & run tests: Ubuntu 16.04 (gcc 5.4.0):
  - standard (implies -std=c11):
    $ cmake -DOPENSSL_NO_ASM=1 .
    $ make all && make run_tests
    PASS
    [100%] Built target run_tests
  - forced -std=gnu99:
    $ cmake -DOPENSSL_NO_ASM=1 -DFORCE_C_STD_FLAGS=-std=gnu99 .
    $ make all && make run_tests
    PASS
    [100%] Built target run_tests